dom/serviceworkers/test/test_offline_localhost_fetch.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1843001</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<iframe></iframe>
<script>

const basePath = window.location.pathname.replace(/\/[^\/]*$/, '/');
const localhostBaseURL = "http://127.0.0.1:8888" + basePath;

  var iframe;
  function runTest() {
    iframe = document.querySelector("iframe");
    iframe.src = localhostBaseURL + "offline_fetch/register.html";
    var ios;
    window.onmessage = function(e) {
      if (e.data.status == "ok") {
        ok(e.data.result, e.data.message);
      } else if (e.data.status == "registrationdone") {
        ios = SpecialPowers.Cc["@mozilla.org/network/io-service;1"]
                           .getService(SpecialPowers.Ci.nsIIOService);
        ios.offline = true;
        iframe.src = localhostBaseURL + "offline_fetch/index.html";
      } else if (e.data.status == "done") {
        ios.offline = false;
        iframe.src = localhostBaseURL + "offline_fetch/unregister.html";
      } else if (e.data.status == "unregistrationdone") {
        window.onmessage = null;
        ok(true, "Test finished successfully");
        SimpleTest.finish();
      }
    };
  }

  SimpleTest.waitForExplicitFinish();
  onload = function() {
    SpecialPowers.pushPrefEnv({"set": [
      ["dom.serviceWorkers.exemptFromPerDomainMax", true],
      ["dom.serviceWorkers.enabled", true],
      ["dom.serviceWorkers.testing.enabled", true],
    ]}, runTest);
  };
</script>
</pre>
</body>
</html>
